refactor(coordinator): cleanup and testing for migration (#885)
