+++ contracts/coordinator/src/contract/query.rs
@@ -3,28 +3,13 @@ use error_stack::{Result, ResultExt};
 use itertools::Itertools;
 use service_registry_api::msg::VerifierDetails;
 
+use crate::contract::errors::Error;
 use crate::msg::{ChainContractsKey, ChainContractsResponse, VerifierInfo};
+use crate::state;
 use crate::state::{
     contracts_by_chain, contracts_by_gateway, contracts_by_prover, contracts_by_verifier,
-    load_protocol_contracts, VERIFIER_PROVER_INDEXED_MAP,
+    VERIFIER_PROVER_INDEXED_MAP,
 };
-
-#[derive(thiserror::Error, Debug, PartialEq)]
-pub enum Error {
-    #[error("protocol contracts (e.g. the router) are not registered yet")]
-    ProtocolNotRegistered,
-
-    #[error(
-        "coordinator failed to retrieve verifier details and corresponding provers. service_name: {service_name}, verifier_address: {verifier_address}"
-    )]
-    VerifierDetailsWithProvers {
-        service_name: String,
-        verifier_address: String,
-    },
-
-    #[error("coordinator failed to retreive chain contracts info")]
-    ChainContractsInfo,
-}
 
 pub fn check_verifier_ready_to_unbond(deps: Deps, verifier_address: Addr) -> Result<bool, Error> {
     Ok(!is_verifier_in_any_verifier_set(deps, &verifier_address))
@@ -35,11 +20,11 @@ pub fn verifier_details_with_provers(
     service_name: String,
     verifier_address: Addr,
 ) -> Result<VerifierInfo, Error> {
-    let config =
-        load_protocol_contracts(deps.storage).change_context(Error::ProtocolNotRegistered)?;
+    let protocol =
+        state::protocol_contracts(deps.storage).change_context(Error::ProtocolNotRegistered)?;
 
     let service_registry: service_registry_api::Client =
-        client::ContractClient::new(deps.querier, &config.service_registry).into();
+        client::ContractClient::new(deps.querier, &protocol.service_registry).into();
 
     let verifier_details: VerifierDetails = service_registry
         .verifier(service_name.clone(), verifier_address.to_string())
